+++ src/main/java/org/prebid/server/bidder/HttpBidderRequester.java
@@ -1,12 +1,10 @@
 package org.prebid.server.bidder;
 
-import com.iab.openrtb.request.App;
 import com.iab.openrtb.request.BidRequest;
 import io.netty.channel.ConnectTimeoutException;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
-import io.vertx.core.MultiMap;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.web.RoutingContext;
@@ -22,13 +20,7 @@ import org.prebid.server.bidder.model.HttpRequest;
 import org.prebid.server.bidder.model.HttpResponse;
 import org.prebid.server.bidder.model.Result;
 import org.prebid.server.execution.Timeout;
-import org.prebid.server.proto.openrtb.ext.request.ExtApp;
-import org.prebid.server.proto.openrtb.ext.request.ExtAppPrebid;
-import org.prebid.server.proto.openrtb.ext.request.ExtRequest;
-import org.prebid.server.proto.openrtb.ext.request.ExtRequestPrebid;
-import org.prebid.server.proto.openrtb.ext.request.ExtRequestPrebidChannel;
 import org.prebid.server.proto.openrtb.ext.response.ExtHttpCall;
-import org.prebid.server.util.HttpUtil;
 import org.prebid.server.vertx.http.HttpClient;
 import org.prebid.server.vertx.http.model.HttpClientResponse;
 
@@ -38,7 +30,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.concurrent.TimeoutException;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -56,22 +47,20 @@ public class HttpBidderRequester {
 
     private static final Logger logger = LoggerFactory.getLogger(HttpBidderRequester.class);
 
-    private static final Set<CharSequence> HEADERS_TO_COPY = Collections.singleton(HttpUtil.SEC_GPC_HEADER.toString());
-
     private final HttpClient httpClient;
     private final BidderRequestCompletionTrackerFactory completionTrackerFactory;
     private final BidderErrorNotifier bidderErrorNotifier;
-    private final String pbsVersion;
+    private final HttpBidderRequestEnricher requestEnricher;
 
     public HttpBidderRequester(HttpClient httpClient,
                                BidderRequestCompletionTrackerFactory completionTrackerFactory,
                                BidderErrorNotifier bidderErrorNotifier,
-                               String pbsVersion) {
+                               HttpBidderRequestEnricher requestEnricher) {
 
         this.httpClient = Objects.requireNonNull(httpClient);
         this.completionTrackerFactory = completionTrackerFactoryOrFallback(completionTrackerFactory);
         this.bidderErrorNotifier = Objects.requireNonNull(bidderErrorNotifier);
-        this.pbsVersion = pbsVersion;
+        this.requestEnricher = Objects.requireNonNull(requestEnricher);
     }
 
     /**
@@ -121,67 +110,11 @@ public class HttpBidderRequester {
     private <T> List<HttpRequest<T>> enrichRequests(List<HttpRequest<T>> httpRequests,
                                                     RoutingContext routingContext,
                                                     BidRequest bidRequest) {
-
         return httpRequests.stream().map(httpRequest -> httpRequest.toBuilder()
-                .headers(enrichHeaders(httpRequest.getHeaders(), routingContext.request().headers(), bidRequest))
+                .headers(requestEnricher.enrichHeaders(httpRequest.getHeaders(),
+                        routingContext.request().headers(), bidRequest))
                 .build())
                 .collect(Collectors.toList());
-    }
-
-    private MultiMap enrichHeaders(MultiMap bidderRequestHeaders,
-                                   MultiMap originalRequestHeaders,
-                                   BidRequest bidRequest) {
-        // some bidders has headers on class level, so we create copy to not affect them
-        final MultiMap bidderRequestHeadersCopy = copyMultiMap(bidderRequestHeaders);
-
-        originalRequestHeaders.entries().stream()
-                .filter(entry -> HEADERS_TO_COPY.contains(entry.getKey())
-                        && !bidderRequestHeadersCopy.contains(entry.getKey()))
-                .forEach(entry -> bidderRequestHeadersCopy.add(entry.getKey(), entry.getValue()));
-        addXPrebidHeader(bidderRequestHeadersCopy, bidRequest);
-        return bidderRequestHeadersCopy;
-    }
-
-    private static MultiMap copyMultiMap(MultiMap source) {
-        final MultiMap copiedMultiMap = MultiMap.caseInsensitiveMultiMap();
-        if (source != null && !source.isEmpty()) {
-            source.forEach(entry -> copiedMultiMap.add(entry.getKey(), entry.getValue()));
-        }
-        return copiedMultiMap;
-    }
-
-    private void addXPrebidHeader(MultiMap headers, BidRequest bidRequest) {
-        final String channelRecord = resolveChannelVersionRecord(bidRequest.getExt());
-        final String sdkRecord = resolveSdkVersionRecord(bidRequest.getApp());
-        final String pbsRecord = createNameVersionRecord("pbs-java", pbsVersion);
-        final String value = Stream.of(channelRecord, sdkRecord, pbsRecord)
-                .filter(Objects::nonNull)
-                .collect(Collectors.joining(","));
-        HttpUtil.addHeaderIfValueIsNotEmpty(headers, HttpUtil.X_PREBID_HEADER, value);
-    }
-
-    private static String resolveChannelVersionRecord(ExtRequest extRequest) {
-        final ExtRequestPrebid extPrebid = extRequest != null ? extRequest.getPrebid() : null;
-        final ExtRequestPrebidChannel channel = extPrebid != null ? extPrebid.getChannel() : null;
-        final String channelName = channel != null ? channel.getName() : null;
-        final String channelVersion = channel != null ? channel.getVersion() : null;
-
-        return createNameVersionRecord(channelName, channelVersion);
-    }
-
-    private static String resolveSdkVersionRecord(App app) {
-        final ExtApp extApp = app != null ? app.getExt() : null;
-        final ExtAppPrebid extPrebid = extApp != null ? extApp.getPrebid() : null;
-        final String sdkSource = extPrebid != null ? extPrebid.getSource() : null;
-        final String sdkVersion = extPrebid != null ? extPrebid.getVersion() : null;
-
-        return createNameVersionRecord(sdkSource, sdkVersion);
-    }
-
-    private static String createNameVersionRecord(String name, String version) {
-        return StringUtils.isNotEmpty(name) && StringUtils.isNotEmpty(version)
-                ? String.format("%s/%s", name, version)
-                : null;
     }
 
     private <T> boolean isStoredResponse(List<HttpRequest<T>> httpRequests, String storedResponse, String bidder) {
