+++ src/main/java/org/prebid/server/spring/config/ServiceConfiguration.java
@@ -32,6 +32,7 @@ import org.prebid.server.bidder.BidderCatalog;
 import org.prebid.server.bidder.BidderDeps;
 import org.prebid.server.bidder.BidderErrorNotifier;
 import org.prebid.server.bidder.BidderRequestCompletionTrackerFactory;
+import org.prebid.server.bidder.HttpBidderRequestEnricher;
 import org.prebid.server.bidder.HttpBidderRequester;
 import org.prebid.server.cache.CacheService;
 import org.prebid.server.cache.model.CacheTtl;
@@ -463,12 +464,18 @@ public class ServiceConfiguration {
             HttpClient httpClient,
             @Autowired(required = false) BidderRequestCompletionTrackerFactory bidderRequestCompletionTrackerFactory,
             BidderErrorNotifier bidderErrorNotifier,
-            VersionInfo versionInfo) {
+            HttpBidderRequestEnricher requestEnricher) {
 
         return new HttpBidderRequester(httpClient,
                 bidderRequestCompletionTrackerFactory,
                 bidderErrorNotifier,
-                versionInfo.getVersion());
+                requestEnricher);
+    }
+
+    @Bean
+    HttpBidderRequestEnricher httpBidderRequestEnricher(VersionInfo versionInfo) {
+
+        return new HttpBidderRequestEnricher(versionInfo.getVersion());
     }
 
     @Bean
